+++ addons/medical/XEH_PREP.hpp
@@ -2,9 +2,4 @@
     Part of the TBMod ( https://github.com/TacticalBaconDevs/TBMod )
     Developed by http://tacticalbacon.de
 */
-
-PREP(getBandageTime);
-PREP(getHealTime);
-PREP(getIVTime);
-PREP(getMedicationTime);
-PREP(getStitchTime);
+PREP(calcTreatmentTime);
